finding a prime of the right size was slightly challenging, but is now accomplished.  metadata names should no longer conflict with real names

+++ transactions/Basics.hpp
@@ -82,7 +82,7 @@ namespace myria{
 			|| from  == Level::undef;
 	}
 
-	using Name = unsigned long long;
+	using Name = long int;
 
 	template<Level l>
 	using choose_strong = typename std::integral_constant<bool, runs_with_strong(l)>::type*;
